pyproject.toml
[project]
name = "ica_agent"
version = "0.1.0"
description = "ADK Agent which produces a fact from a given year and an image based on that fact"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk",
    "google-cloud-aiplatform[adk,agent-engines]",
    "pillow",
    "python-dotenv",
]

[tool.setuptools]
packages = ["ica_agent"]
